Consolidate cleanup executable composition
Keep CLI wiring and the assembled QuickstartCleanup class in the executable script so the package contains only focused implementation modules.
